+++ src/test/php/lang/mirrors/unittest/AnnotationTest.class.php
@@ -2,10 +2,16 @@
 
 use lang\mirrors\Annotation;
 use lang\mirrors\TypeMirror;
+use lang\mirrors\unittest\fixture\IdentityFixture;
 use lang\mirrors\parse\Value;
+use lang\mirrors\parse\Member;
+use lang\mirrors\parse\NewInstance;
 
 class AnnotationTest extends \unittest\TestCase {
   private $type;
+
+  public static $FIXTURE = 'static';
+  const FIXTURE = 'constant';
 
   /** @return void */
   public function setUp() {
@@ -25,6 +31,29 @@ class AnnotationTest extends \unittest\TestCase {
   #[@test]
   public function value() {
     $this->assertEquals('Test', (new Annotation($this->type, 'name', new Value('Test')))->value());
+  }
+
+  #[@test]
+  public function class_reference() {
+    $this->assertEquals(self::class, (new Annotation($this->type, 'name', new Member('self', 'class')))->value());
+  }
+
+  #[@test]
+  public function class_constant() {
+    $this->assertEquals(self::FIXTURE, (new Annotation($this->type, 'name', new Member('self', 'FIXTURE')))->value());
+  }
+
+  #[@test]
+  public function static_class_member() {
+    $this->assertEquals(self::$FIXTURE, (new Annotation($this->type, 'name', new Member('self', '$FIXTURE')))->value());
+  }
+
+  #[@test]
+  public function newinstance() {
+    $this->assertEquals(
+      new IdentityFixture('Test'),
+      (new Annotation($this->type, 'name', new NewInstance('IdentityFixture', [new Value('Test')])))->value()
+    );
   }
 
   #[@test]